--- Unit2/DiceRoller/app/src/main/res/values/strings.xml
@@ -1,4 +1,0 @@
-<resources>
-    <string name="app_name">DiceRoller</string>
-    <string name="roll">Roll</string>
-</resources>
