--- Unit2/DiceRoller/gradle/wrapper/gradle-wrapper.properties
@@ -1,6 +1,0 @@
-#Sat Dec 03 16:50:38 KST 2022
-distributionBase=GRADLE_USER_HOME
-distributionUrl=https\://services.gradle.org/distributions/gradle-7.4-bin.zip
-distributionPath=wrapper/dists
-zipStorePath=wrapper/dists
-zipStoreBase=GRADLE_USER_HOME
