+++ Unit2/Pathway3/MyApplication/app/src/main/java/com/android/tiptime/ui/theme/Color.kt
@@ -1,0 +1,8 @@
+package com.android.tiptime.ui.theme
+
+import androidx.compose.ui.graphics.Color
+
+val Purple200 = Color(0xFFBB86FC)
+val Purple500 = Color(0xFF6200EE)
+val Purple700 = Color(0xFF3700B3)
+val Teal200 = Color(0xFF03DAC5)
