+++ Unit2/Pathway3/MyApplication/app/src/main/res/mipmap-anydpi-v26/ic_launcher.xml
