+++ Unit2/Pathway3/MyApplication/app/src/main/res/values/strings.xml
@@ -1,0 +1,6 @@
+<resources>
+    <string name="app_name">TipTime</string>
+    <string name="calculate_tip">Calculate Tip</string>
+    <string name="cost_of_service">Cost of Service</string>
+    <string name="tip_amount">Tip Amount: %s</string>
+</resources>
