+++ Unit2/Pathway3/MyApplication/app/src/main/res/xml/data_extraction_rules.xml
